stargazer_backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.6.0
requests==2.31.0